Transaction box as regular form
No longer with the button inline. I tried several more options with flexbox and bootstrap (i.e. inline-form, align-items-center); all are fine in the browser but unfortunately not in the compiled version. This latest approach behaves the same across both the browser and compiled version and also in different screen sizes.